+++ inbm/Changelog.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - RTC 495463 - Fix existing driver get uninstalled if AOTA update failed using .deb file package
  - RTC 509509 - Fixes a bug on develop branch that prevents reboot.
  - RTC 508711 - [BUG] Fix Thingsboard Docker Stats widget to display data
+ - RTC 509640 - [BUG] INBC Query(swbom) to exit successfully 
 
 ### Added
  - (507873) Added support for Ubuntu 21.10